implemented error handling and fixed surprised emotion
