+++ pages/settings.py
@@ -1,17 +1,19 @@
 import streamlit as st
+
+import config
 import data_handler as dh
-from conifg import possible_diagnoses
-
-
+from config import possible_diagnoses
+from config import non_matching_diagnos_error_message
+from config import getKeyStringList
 
 # CURRENT USER
 
 current_user_uuid = dh.read_selected_user()
 current_user = dh.get(current_user_uuid)
-st.header("Change Current User")
+st.header("Settings")
+st.subheader("Change Current User")
 
 st.write(f"Current User: {current_user.name}")
-
 
 # SELECT EXISTING USERS ##################################
 all_users = dh.get_all_users()
@@ -32,6 +34,17 @@ if st.button("View User Details"):
         st.write("Last Changed:", selected_user.last_changed)
         st.write("Chat History:", selected_user.chat_history)
 
+# MODIFY DIAGNOSIS #######################################
+st.subheader("Modify Diagnosis")
+
+new_diagnosis = st.text_input("New Diagnosis")
+if st.button("Confirm Diagnosis Change"):
+    if new_diagnosis in possible_diagnoses:
+        dh.modify_diagnosis(selected_user.uuid, new_diagnosis)
+    else:
+        st.write(non_matching_diagnos_error_message + getKeyStringList())
+
+
 # CREATE NEW USER #######################################
 st.subheader("Create New User")
 
@@ -39,36 +52,27 @@ _name = st.text_input("Name")
 _diagnosis = st.text_input("Diagnosis")
 if st.button("Confirm Creation"):
     if _name is not "" and _name is not None and _diagnosis is not "" and _diagnosis is not None:
-        new_user = dh.User(
-            name=_name,
-            diagnosis=_diagnosis,
-        )
+        if _diagnosis in possible_diagnoses:
+            new_user = dh.User(
+                name=_name,
+                diagnosis=_diagnosis,
+            )
 
-        dh.put(new_user.uuid, new_user)
-
+            dh.put(new_user.uuid, new_user)
+        else:
+            st.write(non_matching_diagnos_error_message + getKeyStringList())
     else:
         st.write("please provide correct details for the new User")
 
-# DELETE USER #######################################
-st.subheader("Delete User")
-
-uuid_to_delete = st.text_input("UUID of the user to delete")
-if st.button("Confirm Deletion"):
-    dh.delete(uuid_to_delete)
-
-# MODIFY DIAGNOSIS #######################################
-st.subheader("Modify Diagnosis")
-
-uuid_to_change_diagnosis = st.text_input("UUID of the user to change diagnosis")
-new_diagnosis = st.text_input("New Diagnosis")
-if st.button("Confirm Diagnosis Change"):
-    dh.modify_diagnosis(uuid_to_change_diagnosis, new_diagnosis)
+st.subheader("Deletion & Reset")
 
 # RESET CHAT HISTORY #######################################
-st.subheader("Reset Chat History")
 
-uuid_to_reset_chat_history = st.text_input("UUID of the user to reset chat history")
-if st.button("Confirm Reset Chat History"):
-    dh.reset_chat_history(uuid_to_reset_chat_history)
+if st.button("Reset Chat History"):
+    dh.reset_chat_history(selected_user.uuid)
 
 
+# DELETE USER #######################################
+
+if st.button("Delete User"):
+    dh.delete(selected_user.uuid)
